requirements.txt
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et_xmlfile==2.0.0
Faker==22.0.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-documentai==3.7.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==21.2.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
openpyxl==3.1.2
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.8.0
pytesseract==0.3.10
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.6.0
XlsxWriter==3.1.9
